Implementation of Assessment/Collection Performance & Summary APIs

+++ src/main/java/org/gooru/ds/user/processor/userperf/assessments/UserPerfAssessmentsModel.java
@@ -9,7 +9,7 @@ public class UserPerfAssessmentsModel {
     private String id;    
     private String title;
     private String sessionId;
-    private Integer timeSpent;
+    private Long timeSpent;
     private Integer reaction;
     private Double score;
 	public String getId() {
@@ -30,10 +30,10 @@ public class UserPerfAssessmentsModel {
 	public void setSessionId(String sessionId) {
 		this.sessionId = sessionId;
 	}
-	public Integer getTimeSpent() {
+	public Long getTimeSpent() {
 		return timeSpent;
 	}
-	public void setTimeSpent(Integer timeSpent) {
+	public void setTimeSpent(Long timeSpent) {
 		this.timeSpent = timeSpent;
 	}
 	public Integer getReaction() {
