+++ src/main/java/org/gooru/ds/user/processor/userperf/collections/UserPerfCollectionsModelMapper.java
@@ -18,7 +18,7 @@ public class UserPerfCollectionsModelMapper implements ResultSetMapper<UserPerfC
         UserPerfCollectionsModel model = new UserPerfCollectionsModel();
         model.setId(r.getString(MapperFields.COLLECTION_ID));
         model.setTitle(r.getString(MapperFields.COLLECTION_TITLE));        
-        model.setTimeSpent(r.getInt(MapperFields.COLLECTION_TIMESPENT));        
+        model.setTimeSpent(r.getLong(MapperFields.COLLECTION_TIMESPENT));        
         model.setReaction(r.getInt(MapperFields.COLLECTION_REACTION));
         
         return model;
